style(markdown): rename Content -> MarkdownContent
affects: @medly-components/markdown
