--- packages/markdown/src/components/Content/Content.stories.mdx
@@ -1,16 +1,0 @@
-import { Content } from './Content';
-import { Preview, Story, Meta, Props } from '@storybook/addon-docs/blocks';
-
-<Meta title="Core" component={Content} />
-
-# Content
-
-### Basic Content component
-
-<Preview withToolbar>
-    <Story name="Content">{stories.Basic()}</Story>
-</Preview>
-
-### Props
-
-<Props of={Content} />
