--- packages/markdown/src/components/Content/index.ts
@@ -1,3 +1,0 @@
-import { Content } from './Content';
-
-export default Content;
