+++ packages/markdown/src/components/MarkdownContent/index.ts
@@ -1,0 +1,3 @@
+import { MarkdownContent } from './MarkdownContent';
+
+export default MarkdownContent;
