+++ packages/markdown/src/components/index.ts
@@ -1,1 +1,1 @@
-export { default as Content } from './Content';
+export { default as MarkdownContent } from './MarkdownContent';
